--- src/Sulu/Bundle/SecurityBundle/Resources/config/doctrine/BaseUser.orm.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="Sulu\Bundle\SecurityBundle\Entity\BaseUser">
-
-        <id name="id" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-        <field name="username" type="string" column="username" length="60" unique="true"/>
-        <field name="password" type="string" column="password" length="128"/>
-        <field name="locale" type="string" column="locale" length="60"/>
-        <field name="salt" type="string" column="salt" length="60"/>
-        <field name="locked" column="locked" type="boolean">
-            <options>
-                <option name="default">0</option>
-            </options>
-        </field>
-        <field name="enabled" column="enabled" type="boolean">
-            <options>
-                <option name="default">1</option>
-            </options>
-        </field>
-        <field name="lastLogin" column="lastLogin" type="datetime" nullable="true" />
-        <field name="confirmationKey" column="confirmationKey" type="string" length="128" nullable="true" />
-        <field name="passwordResetToken" column="passwordResetToken" type="string" length="80" nullable="true" unique="true" />
-        <field name="passwordResetTokenExpiresAt" column="passwordResetTokenExpiresAt" type="datetime" nullable="true" />
-        <field name="passwordResetTokenEmailsSent" column="passwordResetTokenEmailsSent" type="integer" nullable="true">
-            <options>
-                <option name="default">0</option>
-            </options>
-        </field>
-        <field name="privateKey" type="string" column="privateKey" length="2000" nullable="true"/>
-        <field name="apiKey" type="guid" column="apiKey" length="128" nullable="true"/>
-        <field name="email" column="email" type="string" length="191" unique="true" nullable="true"/>
-    </mapped-superclass>
-
-</doctrine-mapping>
